Separated pdf-specific functions to own module (#17)
