+++ src/helpers.py
@@ -1,3 +1,13 @@
+"""
+QUIZ REPORTS: helpers
+
+authors:
+@markoprodanovic
+
+last edit:
+Monday, April 06, 2020
+"""
+
 from util import shut_down
 import pandas as pd
 import requests
@@ -73,7 +83,7 @@ def download_quiz_report(report_info):
 
         with (open('raw_reports/' + filename, 'wb')) as output:
             output.write(res.content)
-            print(f'\nOutputting Raw Report: {filename}')
+            print(f'Outputting Raw Report: {filename}\n')
 
         df = pd.read_csv(io.StringIO(res.content.decode('utf-8')))
 
@@ -85,20 +95,3 @@ def download_quiz_report(report_info):
 def generate_random_id():
     return''.join([random.choice(string.ascii_letters
                                  + string.digits) for n in range(15)])
-
-
-def draw_my_ruler(pdf):
-    pdf.drawString(100, 810, 'x100')
-    pdf.drawString(200, 810, 'x200')
-    pdf.drawString(300, 810, 'x300')
-    pdf.drawString(400, 810, 'x400')
-    pdf.drawString(500, 810, 'x500')
-
-    pdf.drawString(10, 100, 'y100')
-    pdf.drawString(10, 200, 'y200')
-    pdf.drawString(10, 300, 'y300')
-    pdf.drawString(10, 400, 'y400')
-    pdf.drawString(10, 500, 'y500')
-    pdf.drawString(10, 600, 'y600')
-    pdf.drawString(10, 700, 'y700')
-    pdf.drawString(10, 800, 'y800')
